make competent dataset

+++ mari_press_books_marnii_corpus_to_dataset.py
@@ -1,11 +1,12 @@
-import codecs
 import os
 import re
+from collections import Counter
 from itertools import chain
 from pathlib import Path
-from typing import Iterable
+from typing import Iterable, Literal
 
 import fitz
+from nltk.tokenize import casual_tokenize
 import numpy as np
 from datasets import Dataset
 from tqdm import tqdm
@@ -14,6 +15,8 @@ from mari_vs_russian_classifier import MariVsRussianClassifier
 
 
 class PDFCorpus:
+    NAME: str
+
     TXT_OUTPUT = Path(__file__).parent / 'magazines_and_student_books_txt'
     TXT_OUTPUT.mkdir(exist_ok=True)
 
@@ -51,27 +54,41 @@ class PDFCorpus:
         for filename in tqdm(self._path.iterdir(), desc=self._desc, total=len(os.listdir(self._path))):
             txt_path = (self.TXT_OUTPUT / filename.with_suffix('.txt').name)
             if txt_path.exists():
+                yield txt_path, txt_path.read_text()
                 continue
             yield txt_path, self._read_pdf(filename)
 
-    def __iter__(self) -> Iterable[str]:
+    def __iter__(self) -> Iterable[tuple[str, str, str]]:
         for txt_path, text in self._read_pdfs():
+            if txt_path.exists():
+                yield txt_path.name, text, self.NAME
+                continue
             text = self._postprocess(text)
             text = self._filter_russian_out(text)
             text = self._DASH_BREAK_REGEXP.sub('', text)
             txt_path.write_text(text)
-            yield text
+            yield txt_path.name, text, self.NAME
+
+
+class PDFCorpusClean(PDFCorpus):
+    NAME = 'clean'
+
+
+class PDFCorpusOCR(PDFCorpus):
+    NAME = 'ocr'
 
 
 class PDFCorpusRemap(PDFCorpus):
+    NAME = 'remap'
+
     _SYM_MAP = {
         'Є': 'Ӧ',  # both
         'є': 'ӧ',
-        '™': 'Ӱ',  # meadow
+        '™': 'Ӱ',
         '¢': 'ӱ',
         'Ў': 'Ӱ',
         'ў': 'ӱ',
-        '‰': 'ҥ',
+        '‰': 'ҥ',  # meadow
         'І': 'Ӓ',  # hill
         'і': 'ӓ',
         'Ї': 'Ӹ',
@@ -86,19 +103,58 @@ class PDFCorpusRemap(PDFCorpus):
         return text
 
 
-_PAD = '\x7F'
-codecs.register_error('pad_with_pad_encode', lambda exception: (_PAD.encode(exception.encoding), exception.end))
-codecs.register_error('pad_with_pad_decode', lambda exception: (_PAD, exception.end))
-
-
 class PDFCorpusFixEncoding(PDFCorpusRemap):
+    NAME = 'fix'
+
     def _postprocess(self, text: str) -> str:
-        encoded = text.encode('cp1252', errors='pad_with_pad_encode')
-        decoded = encoded.decode('cp1251', errors='pad_with_pad_decode')
-        text = decoded.replace(_PAD, '')
+        encoded = text.encode('cp1252', errors='replace')
+        decoded = encoded.decode('cp1251', errors='replace')
+        text = re.sub(r'\?\?+?', '', decoded)
 
         text = self._remap(text)
         return text
+
+
+def _is_hill_mari(text: str) -> bool:
+    lowered = text.lower()
+    hill_mari_symbols = {'ӓ', 'ӹ'}
+    counter = Counter(lowered)
+    symbols = all(counter[sym] >= 5 for sym in hill_mari_symbols)
+    vla = 'влӓ ' in text
+    return symbols and vla
+
+
+def _is_meadow_mari(text: str) -> bool:
+    lowered = text.lower()
+    symbols = Counter(lowered)['ҥ'] >= 2
+    vlaks = ['влак ', 'влакын ']
+    return symbols and any(vlak in lowered for vlak in vlaks)
+
+
+def get_lang(text: str) -> Literal['Hill Mari', 'Meadow Mari', 'Hill and Meadow Mari', 'unknown']:
+    hill = _is_hill_mari(text)
+    meadow = _is_meadow_mari(text)
+    if hill and meadow:
+        return 'Hill and Meadow Mari'
+    if hill:
+        return 'Hill Mari'
+    if meadow:
+        return 'Meadow Mari'
+    return 'unknown'
+
+
+def get_year(name: str) -> int:
+    year = re.findall(r'\d\d\d\d', name)
+    if not year:
+        return 0
+    return int(year[0])
+
+
+def get_num_tokens(subset: Iterable[str]) -> int:
+    num_tokens = 0
+    for text in tqdm(subset, desc='Estimating amount of tokens'):
+        num_tokens += len(casual_tokenize(text))
+    return num_tokens
 
 
 if __name__ == '__main__':
@@ -108,11 +164,28 @@ if __name__ == '__main__':
 
     fixed_very_broken = PDFCorpusFixEncoding(main_root / 'requires_encoding_fix', 'Loading broken PDFs')
     fixed_partly_broken = PDFCorpusRemap(main_root / 'requires_remap', 'Loading partially broken PDFs')
-    fixed_super_clean = PDFCorpus(main_root / 'utf8_already', 'Loading clean PDFS')
-    fixed_after_ocr = PDFCorpus(main_root / 'after_ocr', 'Loading OCR PDFs')
+    fixed_super_clean = PDFCorpusClean(main_root / 'utf8_already', 'Loading clean PDFS')
+    fixed_after_ocr = PDFCorpusOCR(main_root / 'after_ocr', 'Loading OCR PDFs')
 
     dataset = Dataset.from_list([
-        {'text': text}
-        for text in chain(fixed_very_broken, fixed_partly_broken, fixed_super_clean, fixed_after_ocr)
+        {
+            'text': text,
+            'lang': get_lang(text),
+            'name': name,
+            'category': category,
+            'year': get_year(name),
+        }
+        for name, text, category in chain(
+            fixed_very_broken, fixed_partly_broken, fixed_super_clean, fixed_after_ocr
+        )
+        if text and not text.isspace()
     ])
+
+    # print('Total tokens:', get_num_tokens(dataset['text']))
+    # print('Hill tokens:', get_num_tokens(entry['text'] for entry in dataset if entry['lang'] == 'Hill Mari'))
+    # print('Meadow tokens:', get_num_tokens(entry['text'] for entry in dataset if entry['lang'] == 'Meadow Mari'))
+    # print('No OCR tokens:', get_num_tokens(entry['text'] for entry in dataset if entry['category'] != 'ocr'))
+    # print('Modern tokens:', get_num_tokens(entry['text'] for entry in dataset if entry['year'] > 1938))
+
     dataset.save_to_disk(Path(__file__).parent / 'corpora' / 'mari-press-students-books-corpus')
+    # dataset.push_to_hub('OneAdder/mari-pdf-corpus')
